+++ spine-unity/Assets/Spine/Editor/spine-unity/Editor/Components/SkeletonRendererInspector.cs
@@ -91,7 +91,7 @@ namespace Spine.Unity.Editor {
 		readonly GUIContent PhysicsPositionInheritanceFactorLabel = new GUIContent("Position",
 			"When set to non-zero, Transform position movement in X and Y direction is applied to skeleton " +
 			"PhysicsConstraints, multiplied by these " +
-			"\nX and Y scale factors to the right. Typical values are " +
+			"\nX and Y scale factors to the right. Typical (X,Y) values are " +
 			"\n(1,1) to apply XY movement normally, " +
 			"\n(2,2) to apply movement with double intensity, " +
 			"\n(1,0) to apply only horizontal movement, or" +